requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
Pillow
boto3
python-multipart